better naming files for testing, adding more tests for urlproc and whitelisting
