+++ tests/test_core_fileproc.py
@@ -1,5 +1,3 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
 import os
 import pytest
 import tempfile
