+++ tests/test_main_utils.py
@@ -2,6 +2,7 @@ import os
 import pytest
 from urlchecker.main.utils import get_tmpdir
 
+
 def test_get_tmpdir(tmp_path):
     """
     test getting a temporary directory (and creating or not)
@@ -18,8 +19,8 @@ def test_get_tmpdir(tmp_path):
     created = get_tmpdir(base_dir)
     if not os.path.exists(created):
         raise AssertionError
-    
+
     # Name should have prefix
     named = get_tmpdir(prefix="tacos", create=False)
-    if not(os.path.basename(named).startswith('tacos')):
+    if not (os.path.basename(named).startswith("tacos")):
         raise AssertionError
